+++ packages/plugin-oas/src/OperationGenerator.ts
@@ -26,6 +26,7 @@ type Context<TOptions, TPluginOptions extends PluginFactoryOptions> = {
    */
   plugin: Plugin<TPluginOptions>
   mode: KubbFile.Mode
+  UNSTABLE_NAMING?: true
 }
 
 export class OperationGenerator<
@@ -112,7 +113,6 @@ export class OperationGenerator<
     } = {},
   ): OperationSchemas {
     const operationId = operation.getOperationId({ friendlyCase: true })
-    const method = operation.method
     const operationName = transformers.pascalCase(operationId)
 
     const resolveKeys = (schema?: SchemaObject) => (schema?.properties ? Object.keys(schema.properties) : undefined)
@@ -127,7 +127,9 @@ export class OperationGenerator<
       const keys = resolveKeys(schema)
 
       return {
-        name: resolveName(transformers.pascalCase(`${operationId} ${name}`)),
+        name: this.context.UNSTABLE_NAMING
+          ? resolveName(transformers.pascalCase(`${operationId} status ${name}`))
+          : resolveName(transformers.pascalCase(`${operationId} ${name}`)),
         description: (operation.getResponseByStatusCode(statusCode) as OasTypes.ResponseObject)?.description,
         schema,
         operation,
@@ -171,7 +173,9 @@ export class OperationGenerator<
         : undefined,
       request: requestSchema
         ? {
-            name: resolveName(transformers.pascalCase(`${operationId} ${method === 'get' ? 'queryRequest' : 'mutationRequest'}`)),
+            name: this.context.UNSTABLE_NAMING
+              ? resolveName(transformers.pascalCase(`${operationId} RequestData`))
+              : resolveName(transformers.pascalCase(`${operationId} ${operation.method === 'get' ? 'queryRequest' : 'mutationRequest'}`)),
             description: (operation.schema.requestBody as OasTypes.RequestBodyObject)?.description,
             operation,
             operationName,
@@ -181,7 +185,9 @@ export class OperationGenerator<
           }
         : undefined,
       response: {
-        name: resolveName(transformers.pascalCase(`${operationId} ${method === 'get' ? 'queryResponse' : 'mutationResponse'}`)),
+        name: this.context.UNSTABLE_NAMING
+          ? resolveName(transformers.pascalCase(`${operationId} ResponseData`))
+          : resolveName(transformers.pascalCase(`${operationId} ${operation.method === 'get' ? 'queryResponse' : 'mutationResponse'}`)),
         operation,
         operationName,
         schema: {